split utility functions for creating test data from database_test
